requirements.txt
python-dotenv==1.0.0
openai==0.27.8
EdgeGPT==0.2.1
aiomysql==0.2.0
pytz==2023.3
aiogram==3.0.0b7
asyncio==3.4.3
youtube-transcript-api==0.6.0
psutil==5.9.5